requirements.txt
pandas
rdflib
nltk
tqdm
bm25s
numpy
python-dotenv
